requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
django-extensions==3.2.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0